requirements.txt
openai
python-dotenv
yfinance
agno
streamlit
fastapi
uvicorn
youtube_transcript_api